requirements.txt
asgiref==3.11.1
Django==5.0.3
sentry-sdk==2.22.0
django-prometheus==2.3.1
django-environ==0.13.0
gunicorn==25.3.0
packaging==26.0
python-dotenv==1.2.2
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
whitenoise==6.12.0
psycopg2-binary==2.9.10
dj-database-url==2.3.0
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
django-ninja==1.6.2
django-cors-headers==4.7.0
django-ratelimit==4.1.0
django-unfold==0.45.0
supabase==2.12.0
